+++ ChangeLog.md
@@ -17,6 +17,8 @@ ChangeLog
 * Changed: Now return HTTP errors when an addressbook-query REPORT is done
   on a uri that's not a vcard. This should help with debugging this common
   mistake.
+* Changed: `PUT` requests with a `Content-Range` header now emit a 400 status
+  instead of 501, as per RFC7231.
 
 2.0.4 (????-??-??)
 ------------------